Created the DriveHandler class and used it in index.ts
